+++ src/main/java/org/polypheny/simpleclient/scenario/graph/queryBuilder/EdgeLabelMatchBuilder.java
@@ -41,7 +41,7 @@ public class EdgeLabelMatchBuilder extends QueryBuilder {
 
 
     public EdgeLabelMatchBuilder( GraphBenchConfig config ) {
-        int randomSeed = config.seed;
+        long randomSeed = config.seed;
         this.random = new Random( randomSeed );
 
         this.highestLabel = config.maxClusterSize - 1;
